Bug 2027693 part 21 - Fix minor issues in PropMap and Shape code. r=iain
`PropMapTypeToString`: the `LinkedPropMap` type is a base class of all non-compact maps.
Use the concrete types instead.

`ForEachObjectFlag`: flags are now `uint32_t` instead of `uint16_t`.

`setShapeSetForAdd`: add an assertion that's already in similar methods.

Differential Revision: https://phabricator.services.mozilla.com/D293018

+++ js/src/vm/Shape.cpp
@@ -1317,8 +1317,8 @@ void Shape::dump(js::JSONPrinter& json) const {
 
 template <typename KnownF, typename UnknownF>
 void ForEachObjectFlag(ObjectFlags flags, KnownF known, UnknownF unknown) {
-  uint16_t raw = flags.toRaw();
-  for (uint16_t i = 1; i; i = i << 1) {
+  uint32_t raw = flags.toRaw();
+  for (uint32_t i = 1; i; i = i << 1) {
     if (!(raw & i)) {
       continue;
     }
